Задание 4 на проверку
