+++ js/marker.js
@@ -1,0 +1,3 @@
+const marker = {};
+
+export {marker};
